requirements.txt
fastapi[all]==0.77.1
pydantic[email]==1.9.0
fastapi-async-sqlalchemy==0.3.11
mangum==0.15.0
uvicorn==0.17.6
aiosqlite==0.17.0
sqlalchemy-utils==0.38.2
loguru==0.6.0